Switch the FFT MPI decomposition library from pfft to heffe, which supports cufft as a gpu accelerated backend.

GPU is 7x faster on vista. Slightly slower for very small runs due to the universal pencil decomposition, but for anything real it is much faster.

+++ libgadget/cooling_uvfluc.cpp
@@ -92,7 +92,7 @@ init_uvf_table(const std::string& UVFluctuationFile, const double BoxSize, const
     if(fabs(TableBoxSize - BoxMpc) > BoxMpc * 1e-5)
         endrun(0, "Wrong UV fluctuation file! %s is for box size %g Mpc/h, but current box is %g Mpc/h\n", UVFluctuationFile.c_str(), TableBoxSize, BoxMpc);
 
-    if(UVF.Nside * UVF.Nside * UVF.Nside != bh.size)
+    if((size_t) UVF.Nside * UVF.Nside * UVF.Nside != bh.size)
         endrun(0, "Corrupt UV Fluctuation table: Nside = %ld, but table is %lu != %ld^3\n", UVF.Nside, bh.size, UVF.Nside);
 
     message(0, "Using NON-UNIFORM UV BG fluctuations from %s. Median reionization redshift is %g\n", UVFluctuationFile.c_str(), ReionRedshift);
